+++ src/data/events.ts
@@ -1,0 +1,18 @@
+const events = [
+  {
+    name: "Event Name",
+    description: "Description",
+    date: "Thursday, May 21st",
+    location: "Zoom",
+    time: "8:00pm - 9:00pm",
+  },
+  {
+    name: "Event Name",
+    description: "Description",
+    date: "Thursday, May 21st",
+    location: "Zoom",
+    time: "8:00pm - 9:00pm",
+  },
+];
+
+export default events;
